Rename Chunk.hpp Chunk.cpp -> ChunkInfo.(h|c)pp
