+++ include/openPMD/ChunkInfo.hpp
